Added more oop  and organised classes into packages
