+++ data/dukeInfo.txt
@@ -1,3 +1,3 @@
 T,1,new task,
-T,0,vibes,
 D,0,hw ,2019-12-02
+D,1,home ,2019-12-02
