+++ src/main/java/duke/Duke.java
@@ -1,0 +1,40 @@
+package duke;
+
+import duke.command.Command;
+
+public class Duke {
+
+    private Storage storage;
+    private TaskList tasks;
+    private Ui ui;
+
+    public Duke(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        tasks = new TaskList(storage.load());
+    }
+
+    public void run() {
+        ui.showWelcomeMessage();
+        boolean isExit = false;
+        while(!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                ui.showLine();
+                Command c = Parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (DukeException e) {
+                ui.showError(e);
+            } finally {
+                ui.showLine();
+            }
+        }
+    }
+
+
+    public static void main(String[] args) {
+        new Duke("./data/dukeInfo.txt").run();
+
+    }
+}
