+++ src/main/java/duke/TaskList.java
@@ -1,0 +1,33 @@
+package duke;
+
+import java.util.ArrayList;
+
+public class TaskList {
+
+    private ArrayList<Task> tasks;
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public ArrayList<Task> getTasks() {
+        return this.tasks;
+    }
+
+    public int getCount() {
+        return this.tasks.size();
+    }
+
+    public Task get(int i) {
+        return this.tasks.get(i);
+    }
+
+    public void add (Task t) {
+        tasks.add(t);
+    }
+
+    public Task remove(int i) {
+        return tasks.remove(i);
+    }
+
+}
