+++ src/main/java/duke/Todo.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 
 public class Todo extends Task{
