+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,20 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ExitCommand extends Command{
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        System.out.println("Bye. Hope to see you again soon!");
+        storage.write(tasks);
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+
+}
